backend/requirements.txt
# CodeFlow Backend Dependencies
# Install with: pip install -r requirements.txt

flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
anthropic>=0.39.0
requests>=2.28.0
gunicorn>=21.0.0   # Production server (Linux/macOS)
